Fix Way-V only playing out of the left channel.
This may be a weird Ryzen-only prefetching bug, if this bug
was present years ago on the machine I had when I added those
prefetch lines, then I should be publicly shamed for not catching
the regression.

Not sure why I added those lines, obviously low-level optimization,
but they can't be adding much performance and 2018 me doesn't
find that to be a very worthwhile optimization anyway.