Re-work worker thread setup code
When this was first written AMD had no SMT, now it's safer to assume that all CPUs
have SMT.  So re-work the code as such, and clean it up a bit in the process.

I do not own a Mac anymore, so that was not tested.  Nor was Windows, but I suspect
it's just fine given it shares the same code paths as Linux aside from counting CPUs
and setting affinity.