Allow double-precision audio at compile time
On a first gen Ryzen, it only adds perhaps 10% overhead at most

An optimization for memory usage could be made by hard coding the wave pool
items to be floats, then creating float* variants of the functions that use
those buffers.  But that is not a trivial change to make.